refactor(crawlers): move common crawler functionality into base abstract class, make ososedki base crawler inherit from that class
