+++ ososedki_dl/crawlers/__init__.py
@@ -2,20 +2,19 @@ from __future__ import annotations
 
 from typing import Type, Union
 
-from ._common import CrawlerContext
-from .base_crawler import BaseCrawler
+from .base_crawler import BaseCrawler, CrawlerContext
 from .ososedki import crawlers as ososedki_crawlers
+from .ososedki_crawler import OsosedkiBaseCrawler
 from .other import crawlers as other_crawlers
-from .simple_crawler import SimpleCrawler
 
-CrawlerType = Union[Type[BaseCrawler], Type[SimpleCrawler]]
-CrawlerInstance = Union[BaseCrawler, SimpleCrawler]
+CrawlerType = Union[Type[OsosedkiBaseCrawler], Type[BaseCrawler]]
+CrawlerInstance = Union[OsosedkiBaseCrawler, BaseCrawler]
 crawlers: list[CrawlerType] = ososedki_crawlers + other_crawlers
 
 __all__: list[str] = [
-    "CrawlerContext",
     "BaseCrawler",
-    "SimpleCrawler",
-    "crawlers",
+    "CrawlerContext",
     "CrawlerInstance",
+    "crawlers",
+    "OsosedkiBaseCrawler",
 ]
