+++ ososedki_dl/crawlers/ososedki/cosplayboobs.py
@@ -1,9 +1,9 @@
 """Downloader for https://cosplayboobs.com"""
 
-from ..base_crawler import BaseCrawler
+from ..ososedki_crawler import OsosedkiBaseCrawler
 
 
-class CosplayBoobsCrawler(BaseCrawler):
+class CosplayBoobsCrawler(OsosedkiBaseCrawler):
     site_url = "https://cosplayboobs.com"
     base_image_path = "/images/a/"
     album_path = "/album/"
