+++ ososedki_dl/crawlers/ososedki/cosplaythots.py
@@ -1,9 +1,9 @@
 """Downloader for https://cosplaythots.com"""
 
-from ..base_crawler import BaseCrawler
+from ..ososedki_crawler import OsosedkiBaseCrawler
 
 
-class CosplayThotsCrawler(BaseCrawler):
+class CosplayThotsCrawler(OsosedkiBaseCrawler):
     site_url = "https://cosplaythots.com"
     base_image_path = "/images/a/"
     album_path = "/p/"
