+++ ososedki_dl/crawlers/ososedki/ocosplay.py
@@ -1,9 +1,9 @@
 """Downloader for https://ocosplay.com"""
 
-from ..base_crawler import BaseCrawler
+from ..ososedki_crawler import OsosedkiBaseCrawler
 
 
-class OCosplayCrawler(BaseCrawler):
+class OCosplayCrawler(OsosedkiBaseCrawler):
     site_url = "https://ocosplay.com"
     base_image_path = "/images/a/"
     album_path = "/g/"
