+++ ososedki_dl/crawlers/ososedki/ososedki.py
@@ -1,9 +1,9 @@
 """Downloader for https://ososedki.com"""
 
-from ..base_crawler import BaseCrawler
+from ..ososedki_crawler import OsosedkiBaseCrawler
 
 
-class OsosedkiCrawler(BaseCrawler):
+class OsosedkiCrawler(OsosedkiBaseCrawler):
     site_url = "https://ososedki.com"
     base_image_path = "/images/a/"
     album_path = "/photos/"
