+++ ososedki_dl/crawlers/ososedki/vipthots.py
@@ -1,9 +1,9 @@
 """Downloader for https://vipthots.com"""
 
-from ..base_crawler import BaseCrawler
+from ..ososedki_crawler import OsosedkiBaseCrawler
 
 
-class VipThotsCrawler(BaseCrawler):
+class VipThotsCrawler(OsosedkiBaseCrawler):
     site_url = "https://vipthots.com"
     base_image_path = "/images/a/"
     album_path = "/p/"
