+++ ososedki_dl/crawlers/other/wildskirts.py
@@ -8,18 +8,16 @@ from typing import TYPE_CHECKING
 from typing_extensions import override
 
 from ...utils import get_final_path
-from .._common import download_media_items, fetch_soup
-from ..simple_crawler import SimpleCrawler
+from ..base_crawler import BaseCrawler
 
 if TYPE_CHECKING:
     from pathlib import Path
 
-    from aiohttp import ClientSession
     from bs4 import BeautifulSoup
     from bs4.element import NavigableString, Tag
 
 
-class WildskirtsCrawler(SimpleCrawler):
+class WildskirtsCrawler(BaseCrawler):
     site_url = "https://wildskirts.com"
     base_photos_url: str = "https://photos.wildskirts.com"
     base_videos_url: str = "https://video.wildskirts.com"
@@ -55,21 +53,19 @@ class WildskirtsCrawler(SimpleCrawler):
         ]
         return images + videos
 
-    async def fetch_media_urls(self, session: ClientSession, url: str) -> list[str]:
+    async def fetch_media_urls(self, url: str) -> list[str]:
         """
         Asynchronously fetches and returns a list of media URLs from the
         specified page URL.
 
         Args:
-            session (ClientSession): An asynchronous HTTP session used to make
-                the request.
             url (str): The URL of the page to extract media URLs from.
 
         Returns:
             list[str]: A list of media URLs found on the page, or an empty list
             if the page could not be fetched or parsed.
         """
-        soup: BeautifulSoup | None = await fetch_soup(session, url)
+        soup: BeautifulSoup | None = await self.fetch_soup(url)
         return self.wildskirts_media_filter(soup) if soup else []
 
     @override
@@ -95,7 +91,7 @@ class WildskirtsCrawler(SimpleCrawler):
 
         profile: str = profile_url.split("/")[-1]
 
-        soup: BeautifulSoup | None = await fetch_soup(self.context.session, profile_url)
+        soup: BeautifulSoup | None = await self.fetch_soup(profile_url)
         if not soup:
             return []
 
@@ -108,7 +104,7 @@ class WildskirtsCrawler(SimpleCrawler):
         urls: list[str] = [f"{profile_url}/{i}" for i in range(1, total_items + 1)]
         # Fetch media URLs concurrently
         media_urls_lists: list[list[str]] = await asyncio.gather(
-            *[self.fetch_media_urls(self.context.session, url) for url in urls]
+            *[self.fetch_media_urls(url) for url in urls]
         )
         # Flatten the list of lists into a single list
         media_urls: list[str] = [url for sublist in media_urls_lists for url in sublist]
@@ -117,10 +113,4 @@ class WildskirtsCrawler(SimpleCrawler):
 
         album_path: Path = get_final_path(self.context.download_path, profile)
 
-        return await download_media_items(
-            self.context.session,
-            media_urls,
-            album_path,
-            self.context.progress,
-            self.context.task,
-        )
+        return await self.download_media_items(media_urls, album_path)
